+++ var/httpd/htdocs/js/Core.Agent.Dashboard.js
@@ -973,6 +973,12 @@ Core.Agent.Dashboard = (function (TargetNS) {
             DashboardTicketWidgetFilter(WidgetContainer);
         }
 
+        // Reinitialize events for Customer Users table on update.
+        // See bug#14737 for more details (https://bugs.otrs.org/show_bug.cgi?id=14737).
+        if (!$.isEmptyObject(Core.Agent.SwitchToCustomerAction)) {
+            Core.Agent.SwitchToCustomerAction.Init();
+        }
+
         Core.UI.InitMasterAction();
     }