+++ CHANGES.md
@@ -1,4 +1,5 @@
 #6.0.22 ????-??-??
+ - 2019-08-21 Fixed bug#[14737](https://bugs.otrs.org/show_bug.cgi?id=14737) - Switch to customer feature not working in AgentCustomerInformationCenter.
  - 2019-08-13 Fixed bug#[14720](https://bugs.otrs.org/show_bug.cgi?id=14720) - TicketStateSet script task fails if target state equals current state.
  - 2019-08-09 Fixed bug#[14677](https://bugs.otrs.org/show_bug.cgi?id=14677) - Undefined field in title of user task activity dialog edit window.
  - 2019-08-08 Fixed bug#[14665](https://bugs.otrs.org/show_bug.cgi?id=14665) - AutoResponseForWebTicket is not honored in follow-up articles.